进销存系统2.0 主要增加功能：1.能从以前的文件中读取信息2.增加修改信息的功能
