+++ level1/p12_warehouse/进销存/进销存/货物件数.txt
@@ -1,0 +1,1 @@
+2
